+++ LRTable/LRRow.h
@@ -4,11 +4,11 @@
 
 struct RowElement
 {
-	RowElement(const std::string & sym, size_t transition, bool shift)
+	RowElement(const std::string & sym, int transition, bool shift)
 		: inputSym(sym), transition(transition), shift(shift)
 	{}
 
-	RowElement(const std::string & sym, size_t transition, Rule rule, bool shift)
+	RowElement(const std::string & sym, int transition, Rule rule, bool shift)
 		: inputSym(sym), transition(transition), rollup(rule), shift(shift)
 	{}
 	std::string inputSym = "";
